Add multi and single player screens with subtitle support
This update introduces MultiPlayerScreen and SinglePlayerScreen, enhancing video playback functionality with features like subtitles, URL input, and local file selection. Additionally, navigation support and a dedicated SubtitleManagementDialog were added. Existing code was refactored to align with these improvements.
