+++ sample/composeApp/src/commonMain/kotlin/sample/app/MultiPlayerScreen.kt
@@ -1,0 +1,181 @@
+package sample.app
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.VolumeOff
+import androidx.compose.material.icons.automirrored.filled.VolumeUp
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import io.github.kdroidfilter.composemediaplayer.*
+
+@Composable
+fun MultiPlayerScreen() {
+    // List of video URLs
+    val videoUrls = listOf(
+        "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4",
+        "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ElephantsDream.mp4",
+        "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerBlazes.mp4",
+        "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerEscapes.mp4"
+    )
+
+    // Create player states for each video
+    val playerStates = videoUrls.map { url ->
+        val state = rememberVideoPlayerState()
+        LaunchedEffect(url) {
+            state.openUri(url)
+        }
+        state
+    }
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        // Header
+        Text(
+            text = "Multi Video Player",
+            style = MaterialTheme.typography.headlineMedium,
+            fontWeight = FontWeight.Bold,
+            color = MaterialTheme.colorScheme.primary,
+            modifier = Modifier.padding(bottom = 16.dp)
+        )
+
+        // List of videos
+        LazyColumn(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.spacedBy(16.dp)
+        ) {
+            items(playerStates.zip(videoUrls)) { (playerState, url) ->
+                MultiVideoItem(playerState = playerState, videoUrl = url)
+            }
+        }
+    }
+}
+
+@Composable
+fun MultiVideoItem(playerState: VideoPlayerState, videoUrl: String) {
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        shape = RoundedCornerShape(12.dp),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+    ) {
+        Column(
+            modifier = Modifier.padding(8.dp)
+        ) {
+            // Video title
+            Text(
+                text = videoUrl.substringAfterLast('/'),
+                style = MaterialTheme.typography.titleMedium,
+                fontWeight = FontWeight.Bold,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                modifier = Modifier.padding(bottom = 8.dp)
+            )
+
+            // Video player
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(200.dp)
+                    .clip(RoundedCornerShape(8.dp))
+                    .background(MaterialTheme.colorScheme.surfaceVariant)
+            ) {
+                VideoPlayerSurface(
+                    playerState = playerState,
+                    modifier = Modifier.fillMaxSize()
+                )
+
+                if (playerState.isLoading) {
+                    Box(
+                        modifier = Modifier
+                            .fillMaxSize()
+                            .background(Color.Transparent),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        CircularProgressIndicator()
+                    }
+                }
+            }
+
+            // Controls
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
+                horizontalArrangement = Arrangement.SpaceEvenly,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                // Play/Pause button
+                FilledIconButton(
+                    onClick = {
+                        if (playerState.isPlaying) playerState.pause() else playerState.play()
+                    },
+                    colors = IconButtonDefaults.filledIconButtonColors(
+                        containerColor = MaterialTheme.colorScheme.primaryContainer
+                    )
+                ) {
+                    Icon(
+                        imageVector = if (playerState.isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
+                        contentDescription = if (playerState.isPlaying) "Pause" else "Play"
+                    )
+                }
+
+                // Stop button
+                FilledIconButton(
+                    onClick = { playerState.stop() },
+                    colors = IconButtonDefaults.filledIconButtonColors(
+                        containerColor = MaterialTheme.colorScheme.errorContainer
+                    )
+                ) {
+                    Icon(Icons.Default.Stop, contentDescription = "Stop")
+                }
+
+                // Volume control
+                Row(
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    IconButton(
+                        onClick = {
+                            val currentVolume = playerState.volume
+                            if (currentVolume > 0f) {
+                                playerState.volume = 0f
+                            } else {
+                                playerState.volume = 1f
+                            }
+                        }
+                    ) {
+                        Icon(
+                            imageVector = if (playerState.volume > 0f)
+                                Icons.AutoMirrored.Filled.VolumeUp
+                            else
+                                Icons.AutoMirrored.Filled.VolumeOff,
+                            contentDescription = "Volume",
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                    Slider(
+                        value = playerState.volume,
+                        onValueChange = { newVolume -> 
+                            playerState.volume = newVolume
+                        },
+                        valueRange = 0f..1f,
+                        modifier = Modifier.width(100.dp)
+                    )
+                }
+            }
+        }
+    }
+}
