+++ sample/composeApp/src/commonMain/kotlin/sample/app/Screen.kt
@@ -1,0 +1,6 @@
+package sample.app
+
+// Define screens for navigation
+enum class Screen {
+    SinglePlayer, MultiPlayer, VideoAttachmentPlayer
+}
